+++ augur/io.py
@@ -77,13 +77,14 @@ def read_metadata(metadata_file, id_columns=("strain", "name"), chunk_size=None)
     ...
     (5, 14)
     (5, 14)
-    (1, 14)
+    (2, 14)
 
     """
     kwargs = {
         "sep": None,
         "engine": "python",
         "skipinitialspace": True,
+        "na_filter": False,
     }
 
     if chunk_size: